modification of the site linked to the bug under symfony 4. Installation of controllers and entities in their corresponding folder. Modification of the configuration linked to this.
