+++ .idea/dictionaries/mickd.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="mickd">
+    <words>
+      <w>michaelgt</w>
+    </words>
+  </dictionary>
+</component>
