+++ config/routes/annotations.yaml
@@ -1,5 +1,5 @@
 controllers:
-    resource: ../../src/AppBundle/Controller/
+    resource: ../../src/Controller/
     type: annotation
 
 kernel:
