+++ config/services.yaml
@@ -24,10 +24,5 @@ services:
         resource: '../src/Controller'
         tags: ['controller.service_arguments']
 
-    #configuration for the operation of controllers create on symfony 3
-    AppBundle\Controller\:
-        resource: '../src/AppBundle/Controller'
-        tags: ['controller.service_arguments']
-
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
