+++ src/Controller/DefaultController.php
@@ -1,14 +1,12 @@
 <?php
 
-namespace AppBundle\Controller;
+namespace App\Controller;
 
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Twig\Environment;
 
-class DefaultController extends AbstractController
+class DefaultController
 {
     /**
      * @Route("/", name="homepage")
@@ -20,7 +18,7 @@ class DefaultController extends AbstractController
      */
     public function indexAction(Environment $twig)
     {
-        $render= $twig->render('default/index.html.twig');
+        $render = $twig->render('default/index.html.twig');
 
         return new Response($render);
     }
