+++ src/Security/LoginFormAuthenticator.php
@@ -2,7 +2,7 @@
 
 namespace App\Security;
 
-use AppBundle\Entity\User;
+use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
